Inline OpmlImport create
Take advantage of Oaken's env-aware setup phase, so we run this in tests once: `db/seeds/test/setup{,**/*}.rb`.

So we can add test specific helpers methods that are automatically defined and usable in tests! This we can add `db/seeds/test/setup/blobs.rb`.

We'll have to manually call `register` since we're not referencing OpmlImport when Oaken is loading.